+++ Courier.sln
@@ -1,37 +1,69 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Courier", "Courier.csproj", "{37FA9824-8CB2-40A3-BCFB-6672CB2B0F2A}"
+﻿Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Courier", "Courier.csproj", "{37FA9824-8CB2-40A3-BCFB-6672CB2B0F2A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Models.UnitTests", "src\Models.UnitTests\Models.UnitTests.csproj", "{83E5491C-87D0-40DA-8289-9EBB43798E2B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Data.Core.UnitTests", "src\Data\Core.UnitTests\Core.UnitTests.csproj", "{47522C7A-88B3-4679-8C9C-8EC21D0BE933}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Models", "src\Models\Models.csproj", "{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Data.Core", "src\Data\Core\Core.csproj", "{BDF21C22-05A9-4845-9173-B0C64AEC2243}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Monolith", "src\Runtimes\Monolith\Monolith.csproj", "{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Data.InMemoryPersistence", "src\Data\InMemoryPersistence\InMemoryPersistence.csproj", "{6E284B97-9AF7-4BD2-BF82-5C4FF2BCCFCC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SMTP", "src\Servers\SMTP\SMTP.csproj", "{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Data.MongoPersistence", "src\Data\MongoPersistence\MongoPersistence.csproj", "{E6C9FB97-179A-45BD-8846-6308913BF64A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Web", "src\Servers\Web\Web.csproj", "{7B681674-DB2C-43D3-B94A-1EB5C964C628}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Models.UnitTests", "src\Models.UnitTests\Models.UnitTests.csproj", "{83E5491C-87D0-40DA-8289-9EBB43798E2B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Models", "src\Models\Models.csproj", "{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Runtimes.Monolith", "src\Runtimes\Monolith\Monolith.csproj", "{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Servers.SMTP", "src\Servers\SMTP\SMTP.csproj", "{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HawkLab.Courier.Servers.Web", "src\Servers\Web\Web.csproj", "{7B681674-DB2C-43D3-B94A-1EB5C964C628}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{37FA9824-8CB2-40A3-BCFB-6672CB2B0F2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{37FA9824-8CB2-40A3-BCFB-6672CB2B0F2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{37FA9824-8CB2-40A3-BCFB-6672CB2B0F2A}.Release|Any CPU.ActiveCfg = Debug|Any CPU
+		{BDF21C22-05A9-4845-9173-B0C64AEC2243}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BDF21C22-05A9-4845-9173-B0C64AEC2243}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BDF21C22-05A9-4845-9173-B0C64AEC2243}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BDF21C22-05A9-4845-9173-B0C64AEC2243}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4110BAEB-B4B5-4B7C-946F-2E7A58D4997C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{805DD3E7-6473-4B41-B5F3-3840BBC55AD7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E284B97-9AF7-4BD2-BF82-5C4FF2BCCFCC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E284B97-9AF7-4BD2-BF82-5C4FF2BCCFCC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E284B97-9AF7-4BD2-BF82-5C4FF2BCCFCC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E284B97-9AF7-4BD2-BF82-5C4FF2BCCFCC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83E5491C-87D0-40DA-8289-9EBB43798E2B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22A64B6A-7D7C-439E-BA44-E6DE0D2BE624}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E6C9FB97-179A-45BD-8846-6308913BF64A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E6C9FB97-179A-45BD-8846-6308913BF64A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E6C9FB97-179A-45BD-8846-6308913BF64A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E6C9FB97-179A-45BD-8846-6308913BF64A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47522C7A-88B3-4679-8C9C-8EC21D0BE933}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47522C7A-88B3-4679-8C9C-8EC21D0BE933}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47522C7A-88B3-4679-8C9C-8EC21D0BE933}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47522C7A-88B3-4679-8C9C-8EC21D0BE933}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7B681674-DB2C-43D3-B94A-1EB5C964C628}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7B681674-DB2C-43D3-B94A-1EB5C964C628}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B681674-DB2C-43D3-B94A-1EB5C964C628}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B681674-DB2C-43D3-B94A-1EB5C964C628}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
