+++ src/Servers/Web/Pages/Threads/Detail.cshtml.cs
@@ -1,15 +1,14 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using HawkLab.Courier.Models;
-using HawkLab.Courier.Models.Repositories;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-
 namespace MyApp.Namespace
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Threading.Tasks;
+    using HawkLab.Data.Core.Persistence;
+    using HawkLab.Data.Core.Types;
+    using Microsoft.AspNetCore.Mvc;
+    using Microsoft.AspNetCore.Mvc.RazorPages;
+
     public class DetailModel : PageModel
     {
         private readonly IThreadRepository threadRepository;
