+++ src/Servers/Web/Web.csproj
@@ -2,6 +2,6 @@
   <Sdk Name="Microsoft.Build.CentralPackageVersions" />
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Models\Models.csproj" />
+    <ProjectReference Include="..\..\Data\Core\Core.csproj" />
   </ItemGroup>
 </Project>
